+++ src/app/chats/chats.component.html
@@ -35,15 +35,6 @@
                 </div>
             </div>
         </div>
-        <!-- Create chat area -->
-        <!-- Full height can be achived with vh-100 -->
-        <!-- <div class="row">
-            <div class="col col-8 offset-2">
-                <button type="button" class="btn btn-sm btn-block btn-dark"
-                data-toggle="modal" data-target="#newChatConfirmModal">
-                + New Chat</button>
-            </div>
-        </div> -->
         <!-- Modal -->
         <div class="modal fade" id="newChatConfirmModal" tabindex="-1" role="dialog" aria-labelledby="newChatConfirmModalLabel" aria-hidden="true">
             <div class="modal-dialog" role="document">
