--- src/app/in-memory-data.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { InMemoryDataService } from './in-memory-data.service';
-
-describe('InMemoryDataService', () => {
-  let service: InMemoryDataService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(InMemoryDataService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
