--- src/app/mock-chats.ts
@@ -1,9 +1,0 @@
-// Import Chats data model
-import { Chat } from './chat';
-
-// Create fake array for mock chats called CHATS, use Chat data model array
-export const CHATS: Chat[] = [
-    // {id: 1, name: 'Custom Chat One', picture: 'https://png.pngtree.com/element_origin_min_pic/17/08/01/b62d08d09275d0d662a769e2e08e1e9c.jpg'},
-    // {id: 2, name: 'Custom Chat Two', picture: 'https://www.goodfreephotos.com/albums/vector-images/corgi-dog-vector-clipart.png'},
-    // {id: 3, name: 'Custom Chat Three', picture: 'https://previews.123rf.com/images/briang77/briang771512/briang77151200718/49536874-cow-vector-cartoon-illustration.jpg'}
-]
