+++ src/app/users/login/login.component.ts
@@ -19,9 +19,6 @@ export class LoginComponent implements OnInit {
 
   onSubmit(email: string, password: string): void {
     this.authService.credentialSignIn( email, password );
-    // this.authService.credentialSignIn( email, password ).subscribe(user => {
-    //   console.log(user);
-    // });
     this.router.navigate(['']);
   }
